requirements.txt
Django==1.9.1
django-hitcount==1.2.1
django-registration-redux==1.3
django-visits==0.1.6
djangoajax==2.3.3
Pillow==3.1.0
whitenoise==2.0.6
